Attempt to fix feature box text alignment
I set constraints on the height of the box and width of the text in order to fix it. Now the learn more links are misaligned and the images are too small but at least the boxes are all the same dimensions and the top feature box's text align with the one below it.